+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - sideEffects: false
 - BadgeContainer: Added pointer-events: none to badge
 - ScrollArea: Added overflow hidden to root to prevent scrollbar bleeding out in Safari
+- Slider: Tooltip positioning inside the viewport